+++ src/common/editor/area.designer.ts
@@ -8,7 +8,7 @@ import { Splitter, Orientation } from '../splitter';
 import { PropertyGrid } from '../propertygrid';
 import { EditorType } from '../value.editors';
 import { DataGrid } from '../datagrid';
-import { copy, formatString, existsSync, capitalize, Cardinal, pinkfishToHTML, stripPinkfish, consolidate, parseTemplate, initEditDropdown, capitalizePinkfish } from '../library';
+import { copy, formatString, existsSync, capitalize, Cardinal, pinkfishToHTML, stripPinkfish, consolidate, parseTemplate, initEditDropdown, capitalizePinkfish, stripQuotes } from '../library';
 const ResizeObserver = require('resize-observer-polyfill');
 const { clipboard, remote } = require('electron');
 const { Menu, dialog } = remote;
@@ -10237,13 +10237,20 @@ export class AreaDesigner extends EditorBase {
         else {
             const base: Room = this.$area.baseRooms[room.type] || new Room(0, 0, 0);
             if (!room.short || room.short.trim().length === 0)
-                this.$roomPreview.long.textContent = base.short;
+                str = base.short;
             else
-                this.$roomPreview.short.textContent = room.short;
+                str = room.short;
+            if (str.startsWith('"') && str.endsWith('"'))
+                str = str.substr(1, str.length - 2);
+            this.$roomPreview.short.textContent = str;
             if (!room.long || room.long.trim().length === 0)
-                this.$roomPreview.long.textContent = base.long;
+                str = base.long;
             else
-                this.$roomPreview.long.textContent = room.long;
+                str = room.long;
+            if (str.startsWith('"') && str.endsWith('"'))
+                str = str.substr(1, str.length - 2);
+            this.$roomPreview.long.textContent = str;
+
             str = this.$roomPreview.long.innerHTML;
             items = [];
             if ((room.baseFlags & RoomBaseFlags.No_Items) === RoomBaseFlags.No_Items && room.items && room.items.length !== 0)
@@ -11170,12 +11177,10 @@ export class AreaDesigner extends EditorBase {
                 data.description = data.description.trim();
             }
             else if (room.long.length !== 0 || base.long.length === 0) {
-                if (!room.long.startsWith('"') && !room.long.endsWith('"'))
+                if (room.long.startsWith('"') && room.long.endsWith('"'))
+                    room.long = room.long.substr(1, room.long.length - 2);
+                else
                     room.long = room.long.replace(/"/g, '\\"');
-                if (room.long.startsWith('"'))
-                    room.long = room.long.substr(1);
-                if (room.long.endsWith('"'))
-                    room.long = room.long.substr(0, room.long.length - 1);
                 if (room.long.length > 70) {
                     data.description = formatString(room.long, 0, 77, ' * ', '');
                     tmp = room.long.substr(0, 66);
@@ -11205,10 +11210,8 @@ export class AreaDesigner extends EditorBase {
                 data.description = data.description.trim();
             }
             else {
-                if (data.description.startsWith('"'))
-                    data.description = data.description.substr(1);
-                if (data.description.endsWith('"'))
-                    data.description = data.description.substr(0, data.description.length - 1);
+                if (data.description.startsWith('"') && data.description.endsWith('"'))
+                    data.description = data.description.substr(1, data.description.length - 2);
                 if (data.description.length > 70)
                     data.description = formatString(data.description, 0, 77, ' * ', '');
                 else
@@ -11838,12 +11841,11 @@ export class AreaDesigner extends EditorBase {
                 data.description = data.description.trim();
             }
             else if (monster.long.length !== 0 || base.long.length === 0) {
-                if (!monster.long.startsWith('"') && !monster.long.endsWith('"'))
+                if (monster.long.startsWith('"') && monster.long.endsWith('"'))
+                    monster.long = monster.long.substr(1, monster.long.length - 2);
+                else
                     monster.long = monster.long.replace(/"/g, '\\"');
-                if (monster.long.startsWith('"'))
-                    monster.long = monster.long.substr(1);
-                if (monster.long.endsWith('"'))
-                    monster.long = monster.long.substr(0, monster.long.length - 1);
+
                 if (monster.long.length > 70) {
                     data.description = formatString(monster.long, 0, 77, ' * ', '');
                     tmp = monster.long.substr(0, 66);
@@ -11873,10 +11875,8 @@ export class AreaDesigner extends EditorBase {
                 data.description = data.description.trim();
             }
             else {
-                if (data.description.startsWith('"'))
-                    data.description = data.description.substr(1);
-                if (data.description.endsWith('"'))
-                    data.description = data.description.substr(0, data.description.length - 1);
+                if (data.description.startsWith('"') && data.description.endsWith('"'))
+                    data.description = data.description.substr(1, data.description.length - 2);
                 if (data.description.length > 70)
                     data.description = formatString(data.description, 0, 77, ' * ', '');
                 else
@@ -12950,10 +12950,8 @@ export class AreaDesigner extends EditorBase {
             obj.nouns = obj.nouns.split(',');
             obj.nouns = obj.nouns.map(w => {
                 w = w.trim();
-                if (!w.startsWith('"'))
-                    w = '"' + w;
-                if (!w.endsWith('"'))
-                    w += '"';
+                if (!w.startsWith('"') && !w.endsWith('"'))
+                    w = '"' + w.replace(/"/g, '\\"') + '"';
                 return w;
             });
             data['create body'] += '   set_nouns(' + obj.nouns.join(', ') + ');\n';
@@ -12962,10 +12960,8 @@ export class AreaDesigner extends EditorBase {
             obj.adjectives = obj.adjectives.split(',');
             obj.adjectives = obj.adjectives.map(w => {
                 w = w.trim();
-                if (!w.startsWith('"'))
-                    w = '"' + w;
-                if (!w.endsWith('"'))
-                    w += '"';
+                if (!w.startsWith('"') && !w.endsWith('"'))
+                    w = '"' + w.replace(/"/g, '\\"') + '"';
                 return w;
             });
             data['create body'] += '   set_adjectives(' + obj.adjectives.join(', ') + ');\n';